Refactor assemble->add
In the object-oriented API called assembler there was a method
assembler->add that adds an ne*ne dense matrix to an existing large
matrix, where ne is small (such as 2*2 or 3*3) and the large matrix
is in the representation corresponding to this O-O instance (dense,
banded, CSR, etc.).  In each of the (currently existing two) instance
methods for add , there is an almost identical doubly nested loop to
iterate over the two dimensions of the small matrix, look up the
index positions in the large matrix, and add to the large matrix.

I don't see any advantage, in either efficiency or abstraction, to
having this loop-nest inside the assembler abstraction rather than
outside.  Instead, there's a disadvantage: in each new matrix
representation (such as when we add CSR), we'll have to repeat this
loop-nest and its verification.

In this commit, therefore, I refactor, by moving this loop outside the
assembler abstraction into the fem_assemble function.  The
assembler->add method will now just take a single scalar value and the
i,j index at which to add it to the large matrix.